requirements.txt
numpy
pandas
pytorch
pytorch-lightning
scikit-learn
